requirements.txt
# Network/SDN related
ryu==4.34
eventlet==0.30.2
networkx==2.6.3
paramiko==2.11.0
netaddr==0.8.0

# Web server related
flask==2.0.2
werkzeug==2.0.2
requests==2.27.1
aiohttp==3.8.1

# ML related
scikit-learn==1.0.2
numpy==1.22.0
pandas==1.3.5
tensorflow==2.7.0
grpcio==1.43.0
grpcio-tools==1.43.0
protobuf==3.19.0
asyncio==3.4.3

# Utilities
pyyaml==6.0
python-dotenv==0.19.2
matplotlib==3.5.1
seaborn==0.11.2
joblib==1.0.1

# Honeypot
# For Glastopf (some may already be installed by above dependencies)
lxml==4.9.1
beautifulsoup4==4.10.0
chardet==4.0.0
sqlalchemy==1.4.27
gevent==21.12.0

# glutton 